+++ UI/NativeApp.cpp
@@ -91,6 +91,7 @@
 #include "Core/FileLoaders/DiskCachingFileLoader.h"
 #include "Core/KeyMap.h"
 #include "Core/Reporting.h"
+#include "Core/RetroAchievements.h"
 #include "Core/SaveState.h"
 #include "Core/Screenshot.h"
 #include "Core/System.h"
@@ -111,7 +112,6 @@
 #include "UI/BackgroundAudio.h"
 #include "UI/ControlMappingScreen.h"
 #include "UI/DiscordIntegration.h"
-#include "UI/RetroAchievements.h"
 #include "UI/EmuScreen.h"
 #include "UI/GameInfoCache.h"
 #include "UI/GPUDriverTestScreen.h"
