+++ UI/RetroAchievementScreens.cpp
@@ -1,5 +1,4 @@
 #include "UI/RetroAchievementScreens.h"
-#include "UI/RetroAchievements.h"
 #include "Common/System/OSD.h"
 #include "Common/System/Request.h"
 #include "Common/UI/View.h"
@@ -9,6 +8,7 @@
 #include "Common/UI/IconCache.h"
 
 #include "Core/Config.h"
+#include "Core/RetroAchievements.h"
 
 void RetroAchievementsListScreen::CreateTabs() {
 	auto ac = GetI18NCategory(I18NCat::ACHIEVEMENTS);
