+++ UI/UI.vcxproj
@@ -59,7 +59,6 @@
     <ClCompile Include="ProfilerDraw.cpp" />
     <ClCompile Include="RemoteISOScreen.cpp" />
     <ClCompile Include="ReportScreen.cpp" />
-    <ClCompile Include="RetroAchievements.cpp" />
     <ClCompile Include="RetroAchievementScreens.cpp" />
     <ClCompile Include="SavedataScreen.cpp" />
     <ClCompile Include="Store.cpp" />
@@ -96,7 +95,6 @@
     <ClInclude Include="ProfilerDraw.h" />
     <ClInclude Include="RemoteISOScreen.h" />
     <ClInclude Include="ReportScreen.h" />
-    <ClInclude Include="RetroAchievements.h" />
     <ClInclude Include="RetroAchievementScreens.h" />
     <ClInclude Include="SavedataScreen.h" />
     <ClInclude Include="Store.h" />
