+++ UI/UI.vcxproj.filters
@@ -86,7 +86,6 @@
     <ClCompile Include="TabbedDialogScreen.cpp">
       <Filter>Screens</Filter>
     </ClCompile>
-    <ClCompile Include="RetroAchievements.cpp" />
     <ClCompile Include="RetroAchievementScreens.cpp">
       <Filter>Screens</Filter>
     </ClCompile>
@@ -176,7 +175,6 @@
     <ClInclude Include="TabbedDialogScreen.h">
       <Filter>Screens</Filter>
     </ClInclude>
-    <ClInclude Include="RetroAchievements.h" />
     <ClInclude Include="RetroAchievementScreens.h">
       <Filter>Screens</Filter>
     </ClInclude>
