+++ UWP/UI_UWP/UI_UWP.vcxproj.filters
@@ -35,7 +35,6 @@
     <ClCompile Include="..\..\UI\AudioCommon.cpp" />
     <ClCompile Include="..\..\UI\CustomButtonMappingScreen.cpp" />
     <ClCompile Include="..\..\UI\TabbedDialogScreen.cpp" />
-    <ClCompile Include="..\..\UI\RetroAchievements.cpp" />
     <ClCompile Include="..\..\UI\RetroAchievementScreens.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -73,7 +72,6 @@
     <ClInclude Include="..\..\UI\AudioCommon.h" />
     <ClInclude Include="..\..\UI\CustomButtonMappingScreen.h" />
     <ClInclude Include="..\..\UI\TabbedDialogScreen.h" />
-    <ClInclude Include="..\..\UI\RetroAchievements.h" />
     <ClInclude Include="..\..\UI\RetroAchievementScreens.h" />
   </ItemGroup>
 </Project>
